refactor: 💡 create file for get-token-contract-root-bucket
